fix: problem 5

+++ list_1/ex05/main.py
@@ -1,4 +1,4 @@
-number: float = float(input("Diigite um número: "))
+number: int = int(input("Diigite um número: "))
 
 if number % 2 == 0:
     print("par")
